fix: retira metodos do widget Speaker e coloca em um controller
